requirements.txt
pyscreenshot~=3.0
Pillow~=9.2.0
colorthief~=0.2.1
pygame~=2.1.2